Fix parsing of deceptive dates
Previously input that looked numeric could throw off the parser, such as

```
date=1821.a.1
```

It would be parsed as an IDENTIFIER, =, NUMBER(1821), IDENTIFIER(.a.1),
EOF.

This is incorrect. A number should consume all non delimiting characters
(and a period does not constitute a delimiter between fields). Since we
can no longer use regex boundary class `\b`, a postive lookahead had to
be used to ensure that any subsequent characters to a number were
delimiters.