Generate JSON API
There is a middle ground in terms of performance and flexibility: using
JSON as an intermediate layer:

```js
const buffer = readFileSync(args[0]);
const parser = await Jomini.initialize();
const out = parser.parseText(buffer, { encoding: "windows1252" }, (q) => q.json());
const save = JSON.parse(out);
const player = save.player;
const prestige = save.countries[player].prestige;
```

The keys of the stringified JSON object are in the order as they appear
in the file, so this makes the JSON approach well suited for parsing
files where the order of object keys matter. The other APIs are
subjected to natively constructed JS objects reordering keys to suit
their fancy. To process the JSON and not lose key order, you'll want to
leverage a streaming JSON parser like
[oboe.js](https://github.com/jimhigson/oboe.js) or
[stream-json](https://github.com/uhop/stream-json).

Interestingly, even though using JSON adds a layer, constructing and
parsing the JSON into a JS object is still 3x faster than when a JS
object is constructed directly. This must be a testament to how tuned
browser JSON parsers are.

The JSON format does not change how dates are encoded, so dates are
written into the JSON exactly as they appear in the original file.

The biggest downside is that this adds 20KB to the gzipped payload as we
now have to include a JSON library. This seems tolerable.