Add disambiguation option to duplicate keys in JSON output
Previously duplicate keys:

```
core="AAA"
core="BBB"
```

would be coalesced into an array in the JSON output. This is suboptimal
for users that need to distinguish between duplicated keys and / or the
order that keys occur in. The fix is to add a `disambiguate` where one
specifies whether they want duplicated keys in their JSON (and land in
the grey area of validity) or transform objects into an array of key
value tuples.